+++ src/service/result/query.ts
@@ -1,3 +1,4 @@
+import { ColumnMeta, FieldInfo } from "@/common/typeDef";
 import { Util } from "@/common/util";
 import { EsRequest } from "@/model/es/esRequest";
 import { ServiceManager } from "@/service/serviceManager";
@@ -39,7 +40,7 @@ export class QueryPage {
             eventHandler: async (handler) => {
                 handler.on("init", () => {
                     if (queryParam.res?.table) {
-                        handler.panel.title = `${queryParam.res.table}@${dbOption.schema}`
+                        handler.panel.title = queryParam.res.table;
                     }
                     queryParam.res.transId = Trans.transId;
                     queryParam.res.viewId = queryParam.queryOption?.viewId;
@@ -97,7 +98,9 @@ export class QueryPage {
             case MessageType.DATA:
                 if (queryParam.connection.dbType == DatabaseType.ES) {
                     await this.loadEsColumnList(queryParam);
-                } else {
+                }else if (queryParam.connection.dbType == DatabaseType.MONGO_DB) {
+                    await this.loadMongoColumnList(queryParam);
+                }  else {
                     await this.loadColumnList(queryParam);
                 }
                 break;
@@ -150,6 +153,15 @@ export class QueryPage {
         queryParam.res.columnList = queryParam.res.fields.slice(4) as any[]
     }
 
+    private static async loadMongoColumnList(queryParam: QueryParam<DataResponse>) {
+        const parse = queryParam.res.sql.match(/db\('(.+?)'\)\.collection\('(.+?)'\)/);
+        queryParam.res.database = parse[1]
+        queryParam.res.table = parse[2]
+        queryParam.res.primaryKey = '_id'
+        queryParam.res.tableCount = 1
+        queryParam.res.columnList = queryParam.res.fields as any[]
+    }
+
     private static async loadColumnList(queryParam: QueryParam<DataResponse>) {
         // fix null point on result view
         queryParam.res.columnList = []
@@ -168,17 +180,28 @@ export class QueryPage {
             database = fields[0].schema || fields[0].db;
         }
 
+        if (queryParam.connection.dbType == DatabaseType.MSSQL && tableName.indexOf(".") != -1) {
+            tableName = tableName.split(".")[1]
+        }
+
         const tableNode = queryParam.connection.getByRegion(tableName)
         if (tableNode) {
             let primaryKey: string;
+            let primaryKeyList = [];
             const columnList = (await tableNode.getChildren()).map((columnNode: ColumnNode) => {
                 if (columnNode.isPrimaryKey) {
                     primaryKey = columnNode.column.name;
+                    primaryKeyList.push(columnNode.column)
                 }
                 return columnNode.column;
             });
             queryParam.res.primaryKey = primaryKey;
             queryParam.res.columnList = columnList;
+            queryParam.res.primaryKeyList = primaryKeyList;
+            // compatible sqlite empty result.
+            if (queryParam.res.fields.length == 0) {
+                queryParam.res.fields = columnList as any as FieldInfo[];
+            }
         }
         queryParam.res.tableCount = sqlList.length;
         queryParam.res.table = tableName;