+++ src/vue/result/App.vue
@@ -74,7 +74,7 @@
       </ux-table-column>
     </ux-grid>
     <!-- table result -->
-    <EditDialog ref="editor" :dbType="result.dbType" :table="result.table" :primaryKey="result.primaryKey" :columnList="result.columnList" @execute="execute" />
+    <EditDialog ref="editor" :dbType="result.dbType" :database="result.database" :table="result.table" :primaryKey="result.primaryKey" :primaryKeyList="result.primaryKeyList" :columnList="result.columnList" @execute="execute" />
     <ExportDialog :visible.sync="exportOption.visible" @exportHandle="confirmExport" />
   </div>
 </template>
@@ -107,6 +107,7 @@ export default {
         sql: "",
         primaryKey: null,
         columnList: null,
+        primaryKeyList: null,
         database: null,
         table: null,
         tableCount: null,
@@ -207,7 +208,7 @@ export default {
     window.addEventListener("message", ({ data }) => {
       if (!data) return;
       const response = data.content;
-      console.log(response);
+      console.log(data);
       this.table.loading = false;
       switch (data.type) {
         case "EXPORT_DONE":
@@ -234,7 +235,7 @@ export default {
           handlerCommon(response);
           this.info.error = false;
           this.info.needRefresh = false;
-          if(response.message.indexOf("AffectedRows")!=-1){
+          if(response.message.indexOf("AffectedRows")!=-1 || response.isInsert){
             this.refresh()
           }
           break;
@@ -507,7 +508,10 @@ export default {
                     )
                     .join("\n")}`
                 : `DELETE /${this.result.table}/_doc/${checkboxRecords[0]}`;
-          } else {
+          } else if (this.result.dbType == "MongoDB") {
+            deleteSql = `db('${this.result.database}').collection("${this.result.table}")
+              .deleteMany({_id:{$in:[${checkboxRecords.join(",")}]}})`;
+          }else {
             deleteSql =
               checkboxRecords.length > 1
                 ? `DELETE FROM ${this.result.table} WHERE ${